+++ src/CategoryBoards/ToDoList.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { Categories, categoryState, toDoSelector } from '../atoms';
-import CreateToDo from './CreateToDo';
-import ToDo from './ToDo';
+import { categoriesState, toDoSelector } from '../atoms';
+// import CreateToDo from './CreateToDo';
+// import ToDo from './ToDo';
 
 function ToDoList() {
   const toDos = useRecoilValue(toDoSelector);
-  const [category, setCategory] = useRecoilState(categoryState);
+  // const [category, setCategory] = useRecoilState(categoryState);
+  const [categories, setCategory] = useRecoilState(categoriesState);
   const onInput = (event: React.FormEvent<HTMLSelectElement>) => {
     setCategory(event.currentTarget.value as any);
   };
@@ -15,7 +16,8 @@ function ToDoList() {
     <div>
       <h1>To Dos</h1>
       <hr />
-      <select value={category} onInput={onInput}>
+
+      {/* <select value={category} onInput={onInput}>
         <option value={Categories.TO_DO}>To Do</option>
         <option value={Categories.DOING}>Doing</option>
         <option value={Categories.DONE}>Done</option>
@@ -23,7 +25,7 @@ function ToDoList() {
       <CreateToDo />
       {toDos?.map((toDo) => (
         <ToDo key={toDo.id} {...toDo} />
-      ))}
+      ))} */}
     </div>
   );
 }
